email settings update 2

+++ api/application/controllers/v1/backend/SettingsController.php
@@ -1,7 +1,6 @@
 <?php
 defined('BASEPATH') or exit('No direct script access allowed');
 
-use \Firebase\JWT\JWT;
 use \chriskacerguis\RestServer\RestController;
 
 class SettingsController extends RestController
@@ -40,14 +39,16 @@ class SettingsController extends RestController
 
     public function index_get($id)
     {
-        if (!empty($id)) {
-            $settings = $this->settings_model->get(["id" => $id]);
-            $settings->address_informations = json_decode($settings->address_informations, true);
-            $this->response([
-                'status' => TRUE,
-                'message' => "Site Ayarı Başarıyla Getirildi.",
-                'settings' => $settings
-            ], RestController::HTTP_OK);
+        if ($this->token) {
+            if (!empty($id)) {
+                $settings = $this->settings_model->get(["id" => $id]);
+                $settings->address_informations = json_decode($settings->address_informations, true);
+                $this->response([
+                    'status' => TRUE,
+                    'message' => "Site Ayarı Başarıyla Getirildi.",
+                    'settings' => $settings
+                ], RestController::HTTP_OK);
+            }
         }
         $this->response([
             'status' => FALSE,
@@ -79,11 +80,13 @@ class SettingsController extends RestController
 
     public function rank_put($id)
     {
-        if ($this->settings_model->update(["id" => $id], ["rank" => $this->put('rank', true)])) {
-            $this->response([
-                'status' => TRUE,
-                'message' => "Sıralama Başarıyla Güncellendi."
-            ], RestController::HTTP_OK);
+        if ($this->token) {
+            if ($this->settings_model->update(["id" => $id], ["rank" => $this->put('rank', true)])) {
+                $this->response([
+                    'status' => TRUE,
+                    'message' => "Sıralama Başarıyla Güncellendi."
+                ], RestController::HTTP_OK);
+            }
         }
         $this->response([
             'status' => FALSE,
@@ -93,12 +96,14 @@ class SettingsController extends RestController
 
     public function isactive_put($id)
     {
-        $isActive = boolval($this->put("isActive", true)) === true ? 1 : 0;
-        if ($this->settings_model->update(["id" => $id], ["isActive" => $isActive])) {
-            $this->response([
-                'status' => TRUE,
-                'message' => "Durum Başarıyla Güncellendi."
-            ], RestController::HTTP_OK);
+        if ($this->token) {
+            $isActive = boolval($this->put("isActive", true)) === true ? 1 : 0;
+            if ($this->settings_model->update(["id" => $id], ["isActive" => $isActive])) {
+                $this->response([
+                    'status' => TRUE,
+                    'message' => "Durum Başarıyla Güncellendi."
+                ], RestController::HTTP_OK);
+            }
         }
         $this->response([
             'status' => TRUE,
@@ -108,48 +113,50 @@ class SettingsController extends RestController
 
     public function save_post()
     {
-        $data = $this->post();
-        $logo = upload_picture("logo", "uploads/$this->viewFolder", [], "*");
-        $mobile_logo = upload_picture("mobile_logo", "uploads/$this->viewFolder", [], "*");
-        $favicon = upload_picture("favicon", "uploads/$this->viewFolder", [], "*");
-        if ($logo["success"]) :
-            $data["logo"] = $logo["file_name"];
-        endif;
-        if ($mobile_logo["success"]) :
-            $data["mobile_logo"] = $mobile_logo["file_name"];
-        endif;
-        if ($favicon["success"]) :
-            $data["favicon"] = $favicon["file_name"];
-        endif;
-        if (!empty($data["address_informations"])) :
-            foreach (json_decode($data["address_informations"]) as $adKey => $adValue) :
-                json_decode($data["address_informations"])[$adKey]->map = htmlspecialchars(html_entity_decode(json_decode($data["address_informations"])[$adKey]->map));
-            endforeach;
-            $data["address_informations"] = json_encode($data["address_informations"]);
-        endif;
-        $data["rank"] = $this->settings_model->rowCount() + 1;
-        if ($this->settings_model->add($data)) {
-            $this->response([
-                'status' => TRUE,
-                'message' => "Site Ayarları Başarıyla Kayıt Edildi."
-            ], RestController::HTTP_OK);
+        if ($this->token) {
+            $data = $this->post();
+            $logo = upload_picture("logo", "uploads/$this->viewFolder", [], "*");
+            $mobile_logo = upload_picture("mobile_logo", "uploads/$this->viewFolder", [], "*");
+            $favicon = upload_picture("favicon", "uploads/$this->viewFolder", [], "*");
+            if ($logo["success"]) :
+                $data["logo"] = $logo["file_name"];
+            endif;
+            if ($mobile_logo["success"]) :
+                $data["mobile_logo"] = $mobile_logo["file_name"];
+            endif;
+            if ($favicon["success"]) :
+                $data["favicon"] = $favicon["file_name"];
+            endif;
+            if (!empty($data["address_informations"])) :
+                foreach (json_decode($data["address_informations"]) as $adKey => $adValue) :
+                    json_decode($data["address_informations"])[$adKey]->map = htmlspecialchars(html_entity_decode(json_decode($data["address_informations"])[$adKey]->map));
+                endforeach;
+                $data["address_informations"] = json_encode($data["address_informations"]);
+            endif;
+            $data["rank"] = $this->settings_model->rowCount() + 1;
+            if ($this->settings_model->add($data)) {
+                $this->response([
+                    'status' => TRUE,
+                    'message' => "Site Ayarları Başarıyla Kayıt Edildi."
+                ], RestController::HTTP_OK);
+            }
+            if ($logo["success"]) :
+                if (!is_dir(FCPATH . "uploads/{$this->viewFolder}/{$data["logo"]}") && file_exists(FCPATH . "uploads/{$this->viewFolder}/{$data["logo"]}")) :
+                    unlink(FCPATH . "uploads/{$this->viewFolder}/{$data["logo"]}");
+                endif;
+            endif;
+            if ($mobile_logo["success"]) :
+                if (!is_dir(FCPATH . "uploads/{$this->viewFolder}/{$data["mobile_logo"]}") && file_exists(FCPATH . "uploads/{$this->viewFolder}/{$data["mobile_logo"]}")) :
+                    unlink(FCPATH . "uploads/{$this->viewFolder}/{$data["mobile_logo"]}");
+                endif;
+                $data["mobile_logo"] = $mobile_logo["file_name"];
+            endif;
+            if ($favicon["success"]) :
+                if (!is_dir(FCPATH . "uploads/{$this->viewFolder}/{$data["favicon"]}") && file_exists(FCPATH . "uploads/{$this->viewFolder}/{$data["favicon"]}")) :
+                    unlink(FCPATH . "uploads/{$this->viewFolder}/{$data["favicon"]}");
+                endif;
+            endif;
         }
-        if ($logo["success"]) :
-            if (!is_dir(FCPATH . "uploads/{$this->viewFolder}/{$data["logo"]}") && file_exists(FCPATH . "uploads/{$this->viewFolder}/{$data["logo"]}")) :
-                unlink(FCPATH . "uploads/{$this->viewFolder}/{$data["logo"]}");
-            endif;
-        endif;
-        if ($mobile_logo["success"]) :
-            if (!is_dir(FCPATH . "uploads/{$this->viewFolder}/{$data["mobile_logo"]}") && file_exists(FCPATH . "uploads/{$this->viewFolder}/{$data["mobile_logo"]}")) :
-                unlink(FCPATH . "uploads/{$this->viewFolder}/{$data["mobile_logo"]}");
-            endif;
-            $data["mobile_logo"] = $mobile_logo["file_name"];
-        endif;
-        if ($favicon["success"]) :
-            if (!is_dir(FCPATH . "uploads/{$this->viewFolder}/{$data["favicon"]}") && file_exists(FCPATH . "uploads/{$this->viewFolder}/{$data["favicon"]}")) :
-                unlink(FCPATH . "uploads/{$this->viewFolder}/{$data["favicon"]}");
-            endif;
-        endif;
         $this->response([
             'status' => FALSE,
             'message' => "Site Ayarları Kayıt Edilirken Hata Oluştu."
@@ -158,42 +165,44 @@ class SettingsController extends RestController
 
     public function update_post($id)
     {
-        if (!empty($id)) {
-            $settings = $this->settings_model->get(["id" => $id]);
-            if (!empty($settings)) {
-                $data = $this->post();
-                $logo = upload_picture("logo", "uploads/$this->viewFolder", [], "*");
-                $mobile_logo = upload_picture("mobile_logo", "uploads/$this->viewFolder", [], "*");
-                $favicon = upload_picture("favicon", "uploads/$this->viewFolder", [], "*");
-                if ($logo["success"]) :
-                    $data["logo"] = $logo["file_name"];
-                    if (!is_dir(FCPATH . "uploads/{$this->viewFolder}/{$settings->logo}") && file_exists(FCPATH . "uploads/{$this->viewFolder}/{$settings->logo}")) :
-                        unlink(FCPATH . "uploads/{$this->viewFolder}/{$settings->logo}");
+        if ($this->token) {
+            if (!empty($id)) {
+                $settings = $this->settings_model->get(["id" => $id]);
+                if (!empty($settings)) {
+                    $data = $this->post();
+                    $logo = upload_picture("logo", "uploads/$this->viewFolder", [], "*");
+                    $mobile_logo = upload_picture("mobile_logo", "uploads/$this->viewFolder", [], "*");
+                    $favicon = upload_picture("favicon", "uploads/$this->viewFolder", [], "*");
+                    if ($logo["success"]) :
+                        $data["logo"] = $logo["file_name"];
+                        if (!is_dir(FCPATH . "uploads/{$this->viewFolder}/{$settings->logo}") && file_exists(FCPATH . "uploads/{$this->viewFolder}/{$settings->logo}")) :
+                            unlink(FCPATH . "uploads/{$this->viewFolder}/{$settings->logo}");
+                        endif;
                     endif;
-                endif;
-                if ($mobile_logo["success"]) :
-                    $data["mobile_logo"] = $mobile_logo["file_name"];
-                    if (!is_dir(FCPATH . "uploads/{$this->viewFolder}/{$settings->mobile_logo}") && file_exists(FCPATH . "uploads/{$this->viewFolder}/{$settings->mobile_logo}")) :
-                        unlink(FCPATH . "uploads/{$this->viewFolder}/{$settings->mobile_logo}");
+                    if ($mobile_logo["success"]) :
+                        $data["mobile_logo"] = $mobile_logo["file_name"];
+                        if (!is_dir(FCPATH . "uploads/{$this->viewFolder}/{$settings->mobile_logo}") && file_exists(FCPATH . "uploads/{$this->viewFolder}/{$settings->mobile_logo}")) :
+                            unlink(FCPATH . "uploads/{$this->viewFolder}/{$settings->mobile_logo}");
+                        endif;
                     endif;
-                endif;
-                if ($favicon["success"]) :
-                    $data["favicon"] = $favicon["file_name"];
-                    if (!is_dir(FCPATH . "uploads/{$this->viewFolder}/{$settings->favicon}") && file_exists(FCPATH . "uploads/{$this->viewFolder}/{$settings->favicon}")) :
-                        unlink(FCPATH . "uploads/{$this->viewFolder}/{$settings->favicon}");
+                    if ($favicon["success"]) :
+                        $data["favicon"] = $favicon["file_name"];
+                        if (!is_dir(FCPATH . "uploads/{$this->viewFolder}/{$settings->favicon}") && file_exists(FCPATH . "uploads/{$this->viewFolder}/{$settings->favicon}")) :
+                            unlink(FCPATH . "uploads/{$this->viewFolder}/{$settings->favicon}");
+                        endif;
                     endif;
-                endif;
-                if (!empty($data["address_informations"])) :
-                    foreach (json_decode($data["address_informations"]) as $adKey => $adValue) :
-                        json_decode($data["address_informations"])[$adKey]->map = htmlspecialchars(html_entity_decode(json_decode($data["address_informations"])[$adKey]->map));
-                    endforeach;
-                    $data["address_informations"] = json_encode($data["address_informations"]);
-                endif;
-                if ($this->settings_model->update(["id" => $id], $data)) {
-                    $this->response([
-                        'status' => TRUE,
-                        'message' => "Site Ayarları Başarıyla Güncellendi."
-                    ], RestController::HTTP_OK);
+                    if (!empty($data["address_informations"])) :
+                        foreach (json_decode($data["address_informations"]) as $adKey => $adValue) :
+                            json_decode($data["address_informations"])[$adKey]->map = htmlspecialchars(html_entity_decode(json_decode($data["address_informations"])[$adKey]->map));
+                        endforeach;
+                        $data["address_informations"] = json_encode($data["address_informations"]);
+                    endif;
+                    if ($this->settings_model->update(["id" => $id], $data)) {
+                        $this->response([
+                            'status' => TRUE,
+                            'message' => "Site Ayarları Başarıyla Güncellendi."
+                        ], RestController::HTTP_OK);
+                    }
                 }
             }
         }
@@ -205,21 +214,23 @@ class SettingsController extends RestController
 
     public function delete_delete($id)
     {
-        $settings = $this->settings_model->get(["id" => $id]);
-        if ($this->settings_model->delete(["id" => $id])) {
-            if (!is_dir(FCPATH . "uploads/{$this->viewFolder}/{$settings->logo}") && file_exists(FCPATH . "uploads/{$this->viewFolder}/{$settings->logo}")) :
-                unlink(FCPATH . "uploads/{$this->viewFolder}/{$settings->logo}");
-            endif;
-            if (!is_dir(FCPATH . "uploads/{$this->viewFolder}/{$settings->mobile_logo}") && file_exists(FCPATH . "uploads/{$this->viewFolder}/{$settings->mobile_logo}")) :
-                unlink(FCPATH . "uploads/{$this->viewFolder}/{$settings->mobile_logo}");
-            endif;
-            if (!is_dir(FCPATH . "uploads/{$this->viewFolder}/{$settings->favicon}") && file_exists(FCPATH . "uploads/{$this->viewFolder}/{$settings->favicon}")) :
-                unlink(FCPATH . "uploads/{$this->viewFolder}/{$settings->favicon}");
-            endif;
-            $this->response([
-                'status' => TRUE,
-                'message' => "Site Ayarları Başarıyla Silindi."
-            ], RestController::HTTP_OK);
+        if ($this->token) {
+            $settings = $this->settings_model->get(["id" => $id]);
+            if ($this->settings_model->delete(["id" => $id])) {
+                if (!is_dir(FCPATH . "uploads/{$this->viewFolder}/{$settings->logo}") && file_exists(FCPATH . "uploads/{$this->viewFolder}/{$settings->logo}")) :
+                    unlink(FCPATH . "uploads/{$this->viewFolder}/{$settings->logo}");
+                endif;
+                if (!is_dir(FCPATH . "uploads/{$this->viewFolder}/{$settings->mobile_logo}") && file_exists(FCPATH . "uploads/{$this->viewFolder}/{$settings->mobile_logo}")) :
+                    unlink(FCPATH . "uploads/{$this->viewFolder}/{$settings->mobile_logo}");
+                endif;
+                if (!is_dir(FCPATH . "uploads/{$this->viewFolder}/{$settings->favicon}") && file_exists(FCPATH . "uploads/{$this->viewFolder}/{$settings->favicon}")) :
+                    unlink(FCPATH . "uploads/{$this->viewFolder}/{$settings->favicon}");
+                endif;
+                $this->response([
+                    'status' => TRUE,
+                    'message' => "Site Ayarları Başarıyla Silindi."
+                ], RestController::HTTP_OK);
+            }
         }
         $this->response([
             'status' => FALSE,
